Remove Competitors perspective
