+++ quickevent/app/quickevent/plugins/Competitors/src/competitorsplugin.h
@@ -4,24 +4,8 @@
 #include <qf/qmlwidgets/framework/plugin.h>
 
 #include <qf/core/utils.h>
-#include <qf/core/utils/table.h>
 
-namespace qf {
-
-namespace core {
-
-namespace model {
-class SqlTableModel;
-}}
-
-namespace qmlwidgets {
-class Action;
-namespace framework {
-class PartWidget;
-class DockWidget;
-}}
-
-}
+namespace qf::qmlwidgets::framework { class PartWidget; class DockWidget; }
 
 namespace Competitors {
 
@@ -37,28 +21,16 @@ public:
 
 	qf::qmlwidgets::framework::PartWidget *partWidget() {return m_partWidget;}
 
-	Q_INVOKABLE QObject* createCompetitorDocument(QObject *parent);
-	Q_INVOKABLE int editCompetitor(int id, int mode);
+	// Q_INVOKABLE QObject* createCompetitorDocument(QObject *parent);
 
-	Q_SIGNAL int editCompetitorOnPunch(int siid);
-	Q_SIGNAL void dbEventNotify(const QString &domain, int connection_id, const QVariant &payload);
-	Q_SIGNAL void competitorEdited(); // used to clear caches with competitors
+	// Q_SIGNAL int editCompetitorOnPunch(int siid);
+	// Q_SIGNAL void competitorEdited(); // used to clear caches with competitors
 
-	Q_SIGNAL void nativeInstalled();
-
-	Q_SLOT void reloadRegistrationsModel();
-	qf::core::model::SqlTableModel* registrationsModel();
-	const qf::core::utils::Table& registrationsTable();
+	// Q_SIGNAL void nativeInstalled();
 
 	Q_SLOT void onInstalled();
 private:
-	void onRegistrationsDockVisibleChanged(bool on = true);
-	void onDbEventNotify(const QString &domain, int connection_id, const QVariant &data);
-private:
 	qf::qmlwidgets::framework::PartWidget *m_partWidget = nullptr;
-	qf::qmlwidgets::framework::DockWidget *m_registrationsDockWidget = nullptr;
-	qf::core::model::SqlTableModel *m_registrationsModel = nullptr;
-	qf::core::utils::Table m_registrationsTable;
 };
 
 }
