+++ quickevent/app/quickevent/plugins/Competitors/src/competitorswidget.cpp
@@ -1,9 +1,7 @@
 #include "competitorswidget.h"
 #include "ui_competitorswidget.h"
-#include "competitorwidget.h"
 #include "stationsbackupmemorywidget.h"
 
-#include "competitordocument.h"
 #include "competitorsplugin.h"
 
 #include <plugins/Event/src/eventplugin.h>
@@ -101,12 +99,8 @@ CompetitorsWidget::CompetitorsWidget(QWidget *parent) :
 	ui->tblCompetitors->setTableModel(m);
 	m_competitorsModel = m;
 
-	connect(ui->tblCompetitors, &qfw::TableView::editRowInExternalEditor, this, &CompetitorsWidget::editCompetitor, Qt::QueuedConnection);
-	connect(ui->tblCompetitors, &qfw::TableView::editSelectedRowsInExternalEditor, this, &CompetitorsWidget::editCompetitors, Qt::QueuedConnection);
-
 	ui->tblCompetitors->setContextMenuPolicy(Qt::CustomContextMenu);
 	connect(ui->tblCompetitors, &qfw::TableView::customContextMenuRequested, this, &CompetitorsWidget::onCustomContextMenuRequest);
-	connect(getPlugin<CompetitorsPlugin>(), &Competitors::CompetitorsPlugin::editCompetitorOnPunch, this, &CompetitorsWidget::editCompetitorOnPunch);
 
 	QMetaObject::invokeMethod(this, "lazyInit", Qt::QueuedConnection);
 }
@@ -147,15 +141,6 @@ void CompetitorsWidget::settleDownInPartWidget(::PartWidget *part_widget)
 			main_tb->addWidget(m_cbxClasses);
 		}
 		lbl->setBuddy(m_cbxClasses);
-	}
-
-	qf::qmlwidgets::Action *act_print = part_widget->menuBar()->actionForPath("print");
-	act_print->setText(tr("&Print"));
-	{
-		auto *a = new qfw::Action(tr("Competitors statistics"));
-		//a->setShortcut("ctrl+L");
-		connect(a, &qfw::Action::triggered, this, &CompetitorsWidget::report_competitorsStatistics);
-		act_print->addActionInto(a);
 	}
 
 	qf::qmlwidgets::Action *act_stations = part_widget->menuBar()->actionForPath("stations");
@@ -219,119 +204,6 @@ void CompetitorsWidget::reload()
 	m_competitorsModel->reload();
 }
 
-void CompetitorsWidget::editCompetitor_helper(const QVariant &id, int mode, int siid)
-{
-	qfLogFuncFrame() << "id:" << id << "mode:" << mode;
-	//qf::core::sql::Transaction transaction;
-	if (m_editCompetitorLock) {
-		qfDebug() << "Another competitor dialog is opened, ignoring editOnPunch..";
-		return;
-	}
-
-	class EditGuard
-	{
-	public:
-		EditGuard(bool &lock) : m_lock(lock) { m_lock = true; }
-		~EditGuard() { m_lock = false; }
-	private:
-		bool &m_lock;
-	};
-	bool ok = false;
-	bool save_and_next = false;
-
-	{
-		EditGuard guard(m_editCompetitorLock);
-		auto *w = new CompetitorWidget();
-		w->setWindowTitle(tr("Edit Competitor"));
-		qfd::Dialog dlg(QDialogButtonBox::Save | QDialogButtonBox::Cancel, this);
-		dlg.setDefaultButton(QDialogButtonBox::Save);
-		if(mode == qf::core::model::DataDocument::ModeInsert || mode == qf::core::model::DataDocument::ModeEdit) {
-			QPushButton *bt_save_and_next = dlg.buttonBox()->addButton(tr("Save and &next"), QDialogButtonBox::AcceptRole);
-			connect(dlg.buttonBox(), &qf::qmlwidgets::DialogButtonBox::clicked, [&save_and_next, bt_save_and_next](QAbstractButton *button) {
-				save_and_next = (button == bt_save_and_next);
-			});
-		}
-		dlg.setCentralWidget(w);
-		w->load(id, mode);
-		auto *doc = qobject_cast<Competitors::CompetitorDocument*>(w->dataController()->document());
-		QF_ASSERT(doc != nullptr, "Document is null!", return);
-		if(mode == qfm::DataDocument::ModeInsert) {
-			if(siid == 0) {
-				bool is_relays = getPlugin<EventPlugin>()->eventConfig()->isRelays();
-				if(!is_relays) {
-					int class_id = m_cbxClasses->currentData().toInt();
-					if(class_id > 0)
-						doc->setValue("competitors.classId", class_id);
-					else
-						doc->setValue("competitors.classId", QVariant());
-				}
-			}
-			else {
-				w->loadFromRegistrations(siid);
-			}
-		}
-		connect(doc, &Competitors::CompetitorDocument::saved, ui->tblCompetitors, &qf::qmlwidgets::TableView::rowExternallySaved, Qt::QueuedConnection);
-		connect(doc, &Competitors::CompetitorDocument::saved, getPlugin<CompetitorsPlugin>(), &Competitors::CompetitorsPlugin::competitorEdited, Qt::QueuedConnection);
-		ok = dlg.exec();
-		//if(ok)
-		//	transaction.commit();
-		//else
-		//	transaction.rollback();
-
-	}
-	if(ok && save_and_next) {
-		QTimer::singleShot(0, [this]() {
-			this->editCompetitor(QVariant(), qf::core::model::DataDocument::ModeInsert);
-		});
-	}
-}
-
-void CompetitorsWidget::editCompetitors(int mode)
-{
-	if(mode == qfm::DataDocument::ModeDelete) {
-		QList<int> sel_rows = ui->tblCompetitors->selectedRowsIndexes();
-		if(sel_rows.count() <= 1)
-			return;
-		if(qfd::MessageBox::askYesNo(this, tr("Really delete all the selected competitors? This action cannot be reverted."), false)) {
-			qfs::Transaction transaction;
-			int n = 0;
-			for(int ix : sel_rows) {
-				int id = ui->tblCompetitors->tableRow(ix).value(ui->tblCompetitors->idColumnName()).toInt();
-				if(id > 0) {
-					Competitors::CompetitorDocument doc;
-					doc.load(id, qfm::DataDocument::ModeDelete);
-					doc.drop();
-					n++;
-				}
-			}
-			if(n > 0) {
-				if(qfd::MessageBox::askYesNo(this, tr("Confirm deletion of %1 competitors.").arg(n), false)) {
-					transaction.commit();
-					ui->tblCompetitors->reload();
-				}
-				else {
-					transaction.rollback();
-				}
-			}
-		}
-	}
-}
-
-void CompetitorsWidget::editCompetitorOnPunch(int siid)
-{
-	qfs::Query q;
-	q.exec("SELECT id FROM competitors WHERE siId=" + QString::number(siid), qfc::Exception::Throw);
-	if(q.next()) {
-		int competitor_id = q.value(0).toInt();
-		if(competitor_id > 0) {
-			editCompetitor_helper(competitor_id, qfm::DataDocument::ModeEdit, 0);
-		}
-	}
-	else {
-		editCompetitor_helper(QVariant(), qfm::DataDocument::ModeInsert, siid);
-	}
-}
-
 void CompetitorsWidget::onCustomContextMenuRequest(const QPoint &pos)
 {
 	qfLogFuncFrame();
@@ -372,60 +244,3 @@ void CompetitorsWidget::onCustomContextMenuRequest(const QPoint &pos)
 	}
 }
 
-void CompetitorsWidget::report_competitorsStatistics()
-{
-	qfLogFuncFrame();
-	int stage_cnt = getPlugin<EventPlugin>()->stageCount();
-
-	qfs::QueryBuilder qb;
-	qb.select2("classes", "id, name").from("classes").orderBy("classes.name");
-	qf::core::model::SqlTableModel m;
-	m.setQueryBuilder(qb);
-	m.reload();
-	qfu::TreeTable tt = m.toTreeTable();
-	tt.setValue("event", getPlugin<EventPlugin>()->eventConfig()->value("event"));
-	for (int stage_id = 1; stage_id <= stage_cnt; ++stage_id) {
-		QString prefix = "e" + QString::number(stage_id) + "_";
-		QString col_runs_count = prefix + "runCount";
-		QString col_map_count = prefix + "mapCount";
-		tt.appendColumn(col_runs_count, QMetaType(QMetaType::Int));
-		tt.appendColumn(col_map_count, QMetaType(QMetaType::Int));
-		{
-			qfs::QueryBuilder qb;
-			qb.select2("classes", "name")
-				.select("COUNT(runs.id) AS runCount")
-				.select("MAX(classdefs.mapCount) as mapCount") // classdefs.mapCount must be in any agregation function in PSQL, MIN can be here as well
-				.from("classes")
-				.joinRestricted("classes.id", "classdefs.classid", "classdefs.stageId={{stage_id}}")
-				.join("classes.id", "competitors.classId")
-				.joinRestricted("competitors.id", "runs.competitorId", "runs.isRunning AND runs.stageId={{stage_id}}")
-				.groupBy("classes.name")
-				.orderBy("classes.name");
-			QVariantMap qpm;
-			qpm["stage_id"] = stage_id;
-			qfs::Query q;
-			q.execThrow(qb.toString(qpm));
-			int i = 0;
-			while(q.next()) {
-				qf::core::utils::TreeTableRow tt_row = tt.row(i);
-				tt_row.setValue(col_runs_count, q.value("runCount"));
-				tt_row.setValue(col_map_count, q.value("mapCount"));
-				tt.setRow(i, tt_row);
-				i++;
-			}
-		}
-	}
-	qfDebug() << tt.toString();
-	QVariantMap props;
-	//props["isBreakAfterEachClass"] = (opts.breakType() != (int)quickevent::gui::ReportOptionsDialog::BreakType::None);
-	//props["isColumnBreak"] = (opts.breakType() == (int)quickevent::gui::ReportOptionsDialog::BreakType::Column);
-	props["stageCount"] = stage_cnt;
-	QString rep_fn = getPlugin<CompetitorsPlugin>()->findReportFile("competitorsStatistics.qml");
-	qf::qmlwidgets::reports::ReportViewWidget::showReport(this
-								, rep_fn
-								, tt.toVariant()
-								, tr("Competitors statistics")
-								, "competitorsStatistics"
-								, props
-								);
-}
