+++ quickevent/app/quickevent/plugins/Event/src/eventplugin.cpp
@@ -3,6 +3,8 @@
 #include "connectionsettings.h"
 #include "eventdialogwidget.h"
 #include "dbschema.h"
+#include "registrationswidget.h"
+#include "lentcardssettingspage.h"
 #include "stagewidget.h"
 #include "stagedocument.h"
 #include "../../Core/src/widgets/appstatusbar.h"
@@ -10,6 +12,10 @@
 #include "services/serviceswidget.h"
 #include "services/emmaclient.h"
 #include "services/qx/qxclientservice.h"
+
+#include <plugins/Core/src/widgets/settingsdialog.h>
+#include <plugins/Event/src/services/oresultsclient.h>
+#include <plugins/Core/src/coreplugin.h>
 
 #include <quickevent/core/og/timems.h>
 
@@ -32,7 +38,6 @@
 #include <qf/core/sql/connection.h>
 #include <qf/core/sql/transaction.h>
 #include <qf/core/utils/fileutils.h>
-#include <plugins/Event/src/services/oresultsclient.h>
 
 #include <QInputDialog>
 #include <QSqlDatabase>
@@ -58,6 +63,8 @@ namespace qfw = qf::qmlwidgets;
 namespace qff = qf::qmlwidgets::framework;
 namespace qfd = qf::qmlwidgets::dialogs;
 namespace qfs = qf::core::sql;
+
+using qff::getPlugin;
 
 namespace Event {
 
@@ -376,6 +383,25 @@ void EventPlugin::onInstalled()
 		//a->setShortcut(QKeySequence("ctrl+shift+R"));
 		fwk->menuBar()->actionForPath("view")->addActionInto(a);
 	}
+	{
+		connect(this, &Event::EventPlugin::eventOpenChanged, this, &EventPlugin::reloadRegistrationsModel);
+
+		{
+			m_registrationsDockWidget = new qff::DockWidget(nullptr);
+			m_registrationsDockWidget->setObjectName("registrationsDockWidget");
+			m_registrationsDockWidget->setWindowTitle(tr("Registrations"));
+			fwk->addDockWidget(Qt::RightDockWidgetArea, m_registrationsDockWidget);
+			m_registrationsDockWidget->hide();
+			connect(m_registrationsDockWidget, &qff::DockWidget::visibilityChanged, this, &EventPlugin::onRegistrationsDockVisibleChanged);
+
+			auto *a = m_registrationsDockWidget->toggleViewAction();
+			//a->setCheckable(true);
+			a->setShortcut(QKeySequence("ctrl+shift+R"));
+			fwk->menuBar()->actionForPath("view")->addActionInto(a);
+		}
+		auto core_plugin = qf::qmlwidgets::framework::getPlugin<Core::CorePlugin>();
+		core_plugin->settingsDialog()->addPage(new LentCardsSettingsPage());
+	}
 }
 
 void EventPlugin::updateWindowTitle()
@@ -554,16 +580,14 @@ void EventPlugin::onDbEvent(const QString &name, QSqlDriver::NotificationSource 
 	}
 }
 
-void EventPlugin::onDbEventNotify(const QString &domain, int connection_id, const QVariant &data)
+void EventPlugin::onRegistrationsDockVisibleChanged(bool on)
 {
-	Q_UNUSED(connection_id)
-	Q_UNUSED(data)
-	if(domain == QLatin1String(Event::EventPlugin::DBEVENT_STAGE_START_CHANGED)) {
-		//int stage_id = data.toInt();
-		clearStageDataCache();
+	if(on && !m_registrationsDockWidget->widget()) {
+		auto *rw = new RegistrationsWidget();
+		m_registrationsDockWidget->setWidget(rw);
+		rw->checkModel();
 	}
 }
-
 void EventPlugin::repairStageStarts(const qf::core::sql::Connection &from_conn, const qf::core::sql::Connection &to_conn)
 {
 	qfs::Query to_q(to_conn);
@@ -1301,6 +1325,70 @@ void EventPlugin::onServiceDockVisibleChanged(bool on)
 		m_servicesDockWidget->setWidget(rw);
 		rw->reload();
 	}
+}
+
+void EventPlugin::onDbEventNotify(const QString &domain, int connection_id, const QVariant &data)
+{
+	Q_UNUSED(connection_id)
+	qfLogFuncFrame() << "domain:" << domain << "payload:" << data;
+	if(domain == QLatin1String(Event::EventPlugin::DBEVENT_STAGE_START_CHANGED)) {
+		//int stage_id = data.toInt();
+		clearStageDataCache();
+	}
+	else if(domain == QLatin1String(Event::EventPlugin::DBEVENT_REGISTRATIONS_IMPORTED)) {
+		reloadRegistrationsModel();
+	}
+	emit dbEventNotify(domain, connection_id, data);
+}
+
+void EventPlugin::reloadRegistrationsModel()
+{
+	qfLogFuncFrame() << "isEventOpen():" << getPlugin<EventPlugin>()->isEventOpen();
+	if(getPlugin<EventPlugin>()->isEventOpen())
+		registrationsModel()->reload();
+	else
+		registrationsModel()->clearRows();
+	// clear registration table to be regenerated when registrationsTable() will be called
+	m_registrationsTable = qf::core::utils::Table();
+}
+
+qf::core::model::SqlTableModel* EventPlugin::registrationsModel()
+{
+	if(!m_registrationsModel) {
+		m_registrationsModel = new qf::core::model::SqlTableModel(this);
+		m_registrationsModel->addColumn("competitorName", tr("Name"));
+		m_registrationsModel->addColumn("registration", tr("Reg"));
+		m_registrationsModel->addColumn("licence", tr("Lic"));
+		m_registrationsModel->addColumn("siId", tr("SI"));
+		//m_registrationsModel->addColumn("fistName");
+		//m_registrationsModel->addColumn("lastName");
+		qfs::QueryBuilder qb;
+		qb.select2("registrations", "firstName, lastName, licence, registration, siId")
+				.select("COALESCE(lastName, '') || ' ' || COALESCE(firstName, '') AS competitorName")
+				.from("registrations")
+				.orderBy("lastName, firstName");
+		m_registrationsModel->setQueryBuilder(qb, false);
+	}
+	return m_registrationsModel;
+}
+
+const qf::core::utils::Table &EventPlugin::registrationsTable()
+{
+	qf::core::model::SqlTableModel *m = registrationsModel();
+	if(m_registrationsTable.isNull() && !m->table().isNull()) {
+		m_registrationsTable = m->table();
+		auto c_nsk = QStringLiteral("competitorNameAscii7");
+		m_registrationsTable.appendColumn(c_nsk, QMetaType::QString);
+		int ix_nsk = m_registrationsTable.fields().fieldIndex(c_nsk);
+		int ix_cname = m_registrationsTable.fields().fieldIndex(QStringLiteral("competitorName"));
+		for (int i = 0; i < m_registrationsTable.rowCount(); ++i) {
+			qf::core::utils::TableRow &row_ref = m_registrationsTable.rowRef(i);
+			QString nsk = row_ref.value(ix_cname).toString();
+			nsk = QString::fromLatin1(qf::core::Collator::toAscii7(QLocale::Czech, nsk, true));
+			row_ref.setValue(ix_nsk, nsk);
+		}
+	}
+	return m_registrationsTable;
 }
 
 }
