+++ quickevent/app/quickevent/plugins/Event/src/findregistrationedit.h
