+++ quickevent/app/quickevent/plugins/Event/src/lentcardssettingspage.h
