+++ quickevent/app/quickevent/plugins/Event/src/registrationswidget.h
