+++ quickevent/app/quickevent/plugins/Event/src/registrationswidget.ui
