+++ quickevent/app/quickevent/plugins/Runs/qml/reports/competitorsStatistics.qml
